pyproject.toml
[project]
name = "happy-badminton"
version = "0.1.0"
description = "羽毛球比赛预测系统 - Badminton Match Prediction System"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Happy-Badminton Team" }
]

keywords = ["badminton", "prediction", "machine-learning", "sports-analytics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# ============================================
# Dependencies
# ============================================
dependencies = [
    # Data Processing
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "openpyxl>=3.1.0",
    "pyarrow>=14.0.0",

    # Machine Learning
    "scikit-learn>=1.3.0",
    "lightgbm>=4.0.0",
    "xgboost>=2.0.0",
    "catboost>=1.2.0",

    # Hyperparameter Optimization
    "optuna>=3.3.0",
    "optuna-integration>=3.3.0",

    # Model Interpretation
    "shap>=0.43.0",
    "numba>=0.56.0",  # Python 3.10+ compatibility

    # Visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.17.0",

    # Configuration & Logging
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",

    # Utilities
    "tqdm>=4.65.0",
    "joblib>=1.3.0",

    # Web API
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "gunicorn>=22.0.0",
]

# ============================================
# Optional Dependencies (Dev)
# ============================================
[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "ipython>=8.0.0",
    "jupyter>=1.0.0",
]

# ============================================
# UV Dependency Groups
# ============================================
[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

# ============================================
# Tool Configurations
# ============================================
[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501", "E402", "N803", "N806", "I001", "F841", "F541", "F401", "E722", "E741"]

[tool.ruff.lint.per-file-ignores]
"scripts/optimize_sota.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

# ============================================
# Hatch Build Configuration
# ============================================
[tool.hatch.build.targets.wheel]
packages = ["src"]

# ============================================
# Build System
# ============================================
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
